twitter fail safe conditions

+++ UI-FlaskReact-AppEngineFlexVM/FlaskApiProvider/blueprints/post_to_twitter.py
@@ -4,6 +4,8 @@ from flask import request
 import requests
 from .utils.env_variables import *
 from .utils.response import success_response, error_response
+
+from .utils.auth import get_user_uuid_from_token
 import os
 posttwitter_bp = Blueprint('posttwitter_bp', __name__)
 
@@ -12,25 +14,45 @@ posttwitter_bp = Blueprint('posttwitter_bp', __name__)
 def posttwitter():
     received_form_response = json.loads(request.data.decode('utf-8'))
     current_date = datetime.utcnow()
-    user_uuid = received_form_response.get("user_uuid", "Error")
+    user_token = received_form_response.get("user_token", None)
+    if user_token is None:
+        print("get_user_devices: No user token in form response")
+        return error_response(
+            message="Please make sure you have added values for all the fields"
+        )
+
+    user_uuid = get_user_uuid_from_token(user_token)
+    if user_uuid is None:
+        print("get_user_devices: No user uuid")
+        return error_response(
+            message="Invalid User: Unauthorized"
+        )
+
     message = received_form_response.get("message","Failed")
     image_src = received_form_response.get("image_url","")
     filename = "temp.png"
-    request_url = requests.get(image_src, stream=True)
-    print(image_src)
-    if request_url.status_code == 200:
-        with open(filename, 'wb') as image:
-            for chunk in request_url:
-                image.write(chunk)
+    try:
+        request_url = requests.get(image_src, stream=True)
+    except:
+        request_url = None
+    if request_url and request_url.status_code == 200:
+            with open(filename, 'wb') as image:
+                for chunk in request_url:
+                    image.write(chunk)
 
+            query = datastore_client.query(kind='Users')
+            query.add_filter('user_uuid', '=', user_uuid)
+            user = list(query.fetch(1))[0]
+            api.update_with_media(filename, '{} from #{}'.format(message,user.get('organization')))
+            os.remove(filename)
+            return success_response(
+                    message="success"
+                )
+    else:
         query = datastore_client.query(kind='Users')
         query.add_filter('user_uuid', '=', user_uuid)
         user = list(query.fetch(1))[0]
-        api.update_with_media(filename, '{} from #{}'.format(message,user.get('organization')))
-        os.remove(filename)
-        return success_response(
-                message="success"
-            )
-    else:
-        return error_response(message="Image could not be retrieved")
+        if message != "":
+            api.update_status('{} from #{}'.format(message, user.get('organization')))
+        return success_response(message="Image could not be retrieved but posted message")
